Test tweaks
